requirements.txt
Django==1.11.2
Pillow==4.1.1
argparse==1.2.1
certifi==2017.04.17
chardet==3.0.4
distribute==0.6.24
django-autofixture==0.12.1
django-bootstrap-form==3.2.1
docopt==0.6.2
idna==2.5
olefile==0.44
pipreqs==0.4.7
pytz==2017.2
requests==2.18.1
sorl-thumbnail==12.4a1
urllib3==1.21.1
wsgiref==0.1.2
yarg==0.1.9
